Use the http collector rather than the https one
As pointed out by @lorenzoPrimi in
measurement-kit/measurement-kit-app-android#10, we currently do not have
a way to tell the test which SSL certificate bundle should be used.

As long as we don't have that, use the http collector such that I can at
least make sure that the application is working.

See also measurement-kit/measurement-kit-android#12.