Sketch out implementation of on_log()
This is the only bit missing to allow people to run the synchronous
NDT test on Android knowing also the speed measured during it.

Actually, beyond what is merely implemented in here, and assuming it
works as it should, the knowledge gathered working on this is quite
useful to move forward also with async tests.

In fact, what we have here is basically:

1. a way (very simple indeed) to tell Swig to allow us to pass a
   generic Java object as a callback and to let such `jobject`
   arrive intact in C++ wrapper code where we can manipulate it

   (yes, yes, the fact that we pass `Object` in rather than a
    type looks like odd, but I'm confident better than this can
    be done with more SWIG-fu tomorrow morning)

2. a way to callback on an object that should work regardless
   of whether the current C++ thread is attached to a Java thread
   and without even knowing the JNI in the first place

3. the understanding this, even if Android should not kill threads
   at random, especially those running native, the code in here does
   not seem to lead to cases where this would be an issue since we
   don't rely upon saved threads or threads being alive at time T

Cc: @AntonioLangiu because what we have here should shed some light
on a number of Android related issues we discussed earlier

Of course, this could still be a failure because I did not understood
some critical aspect and/or because Android's JNI may be different from
how I think a JNI should work in practice. Also, I did not even tried
to compile the code and spent instead 20+ minutes to write this message.