Heavy refactor of Android build system
This diff implements a heavy refactor of the Android build system
originally started by issues with compiling for mips64.

After lots of experiments, I finally nailed down the problem I had
with mips64 as the android-ndk installed with homebrew not being as
reliable as the android-ndk installed using Android studio (eh).

Yet, in the process of understanding the bug, I developed this
improved build system that in my opinion is worth adding to master
as soon as possible. The most nice improvement in here that now
both MK and its JNI wrappers are compiled in here in just one step.

This is great because we don't need to recompile MK for small
changes and publish the corresponding 50+ MiB archive.

(Originally, we had two build steps because we thought it would
have helped the Portolan guys to have precompiled binaries for
their own JNI, but since we agreed with them that we'll add their
JNI to our own JNI, there's no need for that anymore.)

Also, with this change, we're one step closer with running the build
not only from command line but also using Android studio, which I think
is also a great thing to enable more people integrate our library.

Note that this change the library that is build, to be called just
measurement_kit, since now there's no distinction between MK cross
compiled for Android and MK+JNI cross compiled for Android.