s/a/download: replace script-build-unix with script-build-android
I am now using a separate repository for building for Android, which
seems to greatly simplify the toil of cross compiling, because we don't
need anymore to have the same script work in several contexts.

This specific commit removes the script-build-unix submodule and
replaces it with simple `curl` and `shasum` calls to download and
verify the binaries. Note that the sha256 of the binaries is now
committed in the history of the repository.

Also, since all binaries are now in a single tarball, we just
have a single file to download. It used to be like this. I did
change the behaviour. It was creating too much toil.