Update the configure step to script-build-android
Now jni/Android.mk.in is simpler because we don't split
all the different compiled static libraries.

Also, apparently `-latomic` is gone. Instead, welcome `-lz`,
which is a new dependency (a system dependency on Android)
poised to allow us, in the future, to zip uploads.

Because there is a single archive and a single tree, we
don't need to edit jni/Android.mk.in in an as complex way
as before. So, also script/android/configure is leaner.