Build MK 0.10.2-1-android.1
I have done some work on master since the previous release but I'd rather
avoid using it for v0.10.2-1, because that's just a bugfix release for
the OONI for Android app. Better to use the same build process used for
v0.10.1-2, since this is more in line with delivering an hotfix.