Implement reasonable resources management
This diff modifies the build process to include resources into the
assets shipped along with this library.

It also basically rewrites the resource management class. Now that it's
clear to use we don't care about Java (we haven't heard from any Java
developer using MK so far), we'd rather avoid having separate packages.

Note that, to make tests really run on simulators, I needed to add
the permission to use the network. It seems reasonable to me to set
this functionality into the library manifest, but that could also
be done by setting the functionality into the tests manifest. Still,
since the library uses the network, better to say that, no?