mkall_util.h: add diagnostic code for string misencoding (try 3)
Make diagnostic code unconditional. We'll need to remove it for sure
but this should really be invoked this time.

I didn't read the stack trace in depth and assumed that the code
was going to return an exception. Instead it was aborting right into
env->NewStringUTF therefore there was no point in adding actions
after that function call (as we did in try 1 and try 2).