mkall_util.h: add diagnostic code for string misencoding (try 4)
We're clearly starting to have a problem here. While recent changes
enabled faster development, passing through this is clearly the next
bottleneck to be removed. I'll think about this later.

For now, print also the length of the string. It seems the android
logger is truncating it, and that will allow to confirm.