MKResourcesManager.java: better version.json docs
I was confused by the code and thought we were not updating the
version.json file. We're good because of the way in which the file
is being created, but, since I have been confused once, I could
be confused more than once, or someone else could. Hence add a
comment explaining why we're instead doing the right thing.